requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
pydantic-settings>=2.1.0
feedparser>=6.0.10
python-dateutil>=2.8.2
scikit-learn==1.8.dev0
sentence-transformers>=2.2.2
apscheduler>=3.10.4
python-dotenv>=1.0.0
alembic>=1.13.1
beautifulsoup4>=4.12.0 